Simplify middleware
