--- components/idiot/auth/utils.ts
@@ -1,74 +1,0 @@
-import { NextRequest } from 'next/server';
-
-export const generateCookieString = (key: string, value: string, age: string): string =>
-  `${key}=${value}; Domain=${process.env.NEXT_PUBLIC_COOKIE_DOMAIN}; Path=/; Max-Age=${age}; SameSite=strict;`;
-
-export const getQueryParams = (req: NextRequest): any =>
-  req.url.includes('?')
-    ? Object.assign(
-        {},
-        ...req.url
-          .split('?')[1]
-          .split('&')
-          .map((param) => ({ [param.split('=')[0]]: param.split('=')[1] })),
-      )
-    : {};
-
-export const getRequestedURI = (req: NextRequest): string => {
-  return req.url
-    .split('?')[0]
-    .replace(/localhost:\d+/, (process.env.APP_URI || '').replace('https://', '').replace('http://', ''));
-};
-
-export const getJWT = (req: NextRequest) => {
-  const rawJWT = req.cookies.get('jwt')?.value;
-  // Strip any and all 'Bearer 's off of JWT.
-  const jwt = rawJWT?.split(' ')[rawJWT?.split(' ').length - 1] ?? rawJWT ?? '';
-  return jwt;
-};
-export const verifyJWT = async (jwt: string): Promise<Response> => {
-  if (!process.env.SERVERSIDE_AGIXT_SERVER) {
-    process.env.SERVERSIDE_AGIXT_SERVER = ['agixt', 'localhost', 'back-end', 'boilerplate', 'back-end-image'].join(',');
-  }
-  const containerNames = process.env.SERVERSIDE_AGIXT_SERVER.split(',');
-  const responses = {} as any;
-  const authEndpoint = `${process.env.AGIXT_SERVER}/v1/user`;
-  let response;
-  for (const containerName of containerNames) {
-    const testEndpoint = authEndpoint.replace('localhost', containerName);
-    try {
-      response = await fetch(testEndpoint, {
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `${jwt}`,
-        },
-      });
-
-      if (response.status === 200 || [401, 402, 403].includes(response.status)) {
-        if (Object.keys(responses).length > 0) {
-          containerNames.sort((a, b) => {
-            if (a === containerName) {
-              return -1;
-            } else if (b === containerName) {
-              return 1;
-            } else {
-              return 0;
-            }
-          });
-          process.env.SERVERSIDE_AGIXT_SERVER = containerNames.join(',');
-        }
-        return response;
-      } else {
-        responses[testEndpoint] = await response.text();
-        console.error(`Failed to contact server at ${testEndpoint}.`);
-      }
-    } catch (exception) {
-      responses[testEndpoint] = exception;
-    }
-  }
-  console.error('Failed to contact any of the following servers: ', JSON.stringify(responses));
-  for (const key of Object.keys(responses)) {
-    console.error(key, responses[key]);
-  }
-  return new Response();
-};
